Run model forwards under bf16 autocast in CFR search.
Wraps every search-side model call in torch.autocast(bfloat16): both
chance_node_helper sites (were fp16, switched to bf16 to match training),
the parent CFREvaluator's _get_model_policy_probs and _set_model_values_impl,
the FusedSparseCFREvaluator override, and the six call sites in CFRManager.
CFR state (beliefs, reach, regrets, policy, latest_values) stays fp32; bf16
is confined to the MLP forward, with outputs cast back to fp32 at the
boundary before any accumulation or mixing.

+++ src/p2/search/chance_node_helper.py
@@ -305,7 +305,7 @@ class ChanceNodeHelper:
                 board=board_samples_flat,
                 beliefs=belief_features,
             )
-            with torch.autocast(device_type="cuda", dtype=torch.float16):
+            with torch.autocast(device_type="cuda", dtype=torch.bfloat16):
                 hand_values = model(synthetic_features).hand_values
             hand_values = hand_values.to(dtype=dtype).view(
                 B, chunk_len, self.num_players, NUM_HANDS
@@ -461,7 +461,7 @@ class ChanceNodeHelper:
 
         # Note: Technically these are EVs from the model, not CFVs.
         # But reach-weight only changes evenly across the chance node, so we ignore it.
-        with torch.autocast(device_type="cuda", dtype=torch.float16):
+        with torch.autocast(device_type="cuda", dtype=torch.bfloat16):
             hand_values = model(synthetic_features).hand_values
         hand_values = hand_values.to(dtype=dtype)
